ref: refactored startup process into static and dynamic
